fix(projects): stamp project completedAt + integer order qty (round 6)
Two round-6 consistency fixes.

- projects/[id] PATCH stamped the status through but never set completedAt,
  so a project completed via the generic web PATCH kept completedAt = null —
  while both the mobile projects PATCH and the web phase route stamp it.
  Now mirrors them: set completedAt on entering "completed", clear it on
  leaving. (A timestamp stamp, not a transition block.)
- mobile/orders POST line-item qty tightened to z.number().int().positive();
  orderItems.quantity is an integer column, so a fractional qty (e.g. 2.5)
  previously reached Postgres and errored 500 instead of a clean 422. (The
  mobile invoice route keeps fractional qty — invoiceItems.quantity is
  numeric — so the change is order-specific.)

Adds 3 route tests (completedAt stamped on complete, cleared on reopen,
fractional order qty rejected).

+++ prv/apps/web/src/app/api/mobile/orders/route.ts
@@ -12,7 +12,7 @@ export const runtime = "nodejs"
 
 const lineItemSchema = z.object({
   name: z.string().min(1).max(500),
-  qty: z.number().positive(),
+  qty: z.number().int().positive(),
   unitPrice: z.number().min(0),
   vatRate: z.number().min(0).max(100).default(19),
 })
